+++ src/tests/run_tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # Prepare Docker Images
-docker pull ${QL_PREV_IMAGE}
+docker pull orchestracities/quantumleap:${PREV_QL}
 docker build -t orchestracities/quantumleap ../../
 CRATE_VERSION=${PREV_CRATE} docker-compose -f docker-compose-bc.yml pull --ignore-pull-failures
 
@@ -10,8 +10,24 @@ tot=0
 # Launch services with previous CRATE and QL version
 echo "\n"
 echo "Launch services with previous CRATE and QL version"
-CRATE_VERSION=${PREV_CRATE} QL_IMAGE=${QL_PREV_IMAGE} docker-compose -f docker-compose-bc.yml up -d
-sleep 20
+
+CRATE_VERSION=${PREV_CRATE} QL_VERSION=${PREV_QL} docker-compose -f docker-compose-bc.yml up -d
+
+HOST="http://localhost:4200"
+echo "Testing $HOST"
+wait=0
+while [ "$(curl -s -o /dev/null -L -w ''%{http_code}'' $HOST)" != "200" ] && [ $wait -lt 30 ]
+do
+  echo "Waiting for $HOST"
+  sleep 5
+  wait=$((wait+5))
+  echo "Elapsed time: $wait"
+done
+
+if [ $wait -gt 30 ]; then
+  echo "timeout while waiting services to be ready"
+  exit -1
+fi
 
 
 ORION_BC_HOST=`docker inspect -f '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $(docker ps | grep "1026" | awk '{ print $1 }')`
@@ -25,12 +41,28 @@ docker run -ti --rm --network tests_default \
            -e QL_URL="http://$QL_BC_HOST:8668" \
            --entrypoint "" \
            -e USE_FLASK=TRUE \
-           orchestracities/quantumleap:0.8.0 python tests/common.py
+           orchestracities/quantumleap:${PREV_QL} python tests/common.py
 
 # Restart QL on development version and CRATE on current version
-docker-compose stop quantumleap
-CRATE_VERSION=${CRATE_VERSION} QL_IMAGE=orchestracities/quantumleap docker-compose -f docker-compose-bc.yml up -d
-sleep 40
+echo "\n"
+echo "Use current version of ql and crate"
+
+CRATE_VERSION=${CRATE_VERSION} QL_VERSION=latest docker-compose -f docker-compose-bc.yml stop quantumleap crate
+CRATE_VERSION=${CRATE_VERSION} QL_VERSION=latest docker-compose -f docker-compose-bc.yml up -d
+
+wait=0
+while [ "$(curl -s -o /dev/null -L -w ''%{http_code}'' $HOST)" != "200" ] && [ $wait -lt 30 ]
+do
+  echo "Waiting for $HOST"
+  sleep 5
+  wait=$((wait+5))
+  echo "Elapsed time: $wait"
+done
+
+if [ $wait -gt 30 ]; then
+  echo "timeout while waiting services to be ready"
+  exit -1
+fi
 
 # Backwards Compatibility Test
 echo "\n"
@@ -52,4 +84,5 @@ fi
 cd -
 
 docker-compose -f docker-compose-bc.yml down -v
+
 exit ${tot}